+++ DOC/PY/Readme.md
@@ -15,26 +15,17 @@ pinpoint| 1.8.0-RC1(thrift) <br> 2.0+(GRPC)
 
 > Steps
 
-1. Compile pinpointPy module
-```
-$ cd common
-$ mkdir -p build
-$ cd build
-$ cmake -DWITH_TEST_CASE=1 -DCMAKE_BUILD_TYPE=Release .. 
-$ make 
-$ ./test/unittest
-```
-2. Install python virtual environment,refer to 
+1. Install python virtual environment,refer to 
 https://packaging.python.org/guides/installing-using-pip-and-virtual-environments/
 
-3. Come back to the root directory, Install pinpointPy(Python virtualenv is recommend).
+2. Come back to the root directory, Install pinpointPy(Python virtualenv is recommend).
 ```shell
-$ pip install -r requirements.txt 
 $ python setup.py install
-$ nosetests src/PY/test
 ```
 
 ### [How To Use](Example/PY/Readme.md)
+
+
 
 ## Changes
