+++ Example/PHP/demo/yii-demo/Readme.md
@@ -52,5 +52,5 @@
 
 
 ## References
-1. How to install django? [ click me ☚ ](https://docs.djangoproject.com/zh-hans/3.0/intro/install/#install-django)
+1. How to install yii? [ click me ☚ ](https://www.yiiframework.com/doc/guide/2.0/en/start-installation)
 2. What's php_simple_aop? [ click me ☚ ](https://github.com/eeliu/php_simple_aop)
