+++ Example/PY/tornado/Readme.md
@@ -3,7 +3,7 @@
 
 ## Integration pinpoint
 
-> Make sure pinpointPy module has been installed. ([How to Install pinpointPy module](DOC/PY/Readme.md))
+> Make sure pinpointPy module has been installed. ([How to Install pinpointPy module](../../../DOC/PY/Readme.md))
 1. Copy [plugins](tornado/plugins) to your root. Plugins in tornado/plugins are some examples, you can also write your own plugin according to these examples.
 
 2. Use preRequestPlugins(TornadoHTTPRequestPlugins) to bind the requests handdler function.
@@ -18,7 +18,3 @@
         @PinpointCommonPlugin('',__name__)
         async def get_1(self,t = 2):
     ```
-
-## References
-1. How to use tornado? [ click me ☚ ](https://www.tornadoweb.org/en/stable/guide.html)
-2. What's php_simple_aop? [ click me ☚ ](https://github.com/eeliu/php_simple_aop)
